+++ example/lib/localization/application/application_feature_es.dart
@@ -16,4 +16,8 @@ class ApplicationFeatureEs extends ApplicationFeature {
 
   @override
   String get incrementButton => 'Incremento';
+
+  @override
+  String get unTranslatedMessage =>
+      'Hey, i\'m here to showcase messages that are not translated';
 }